refactor: move evaluateBackPressure to abstract base class
